Commented out functions.
comment for function declarations in include/ds/queue.h.
added another function for data rellocation on a queue.

+++ kernel/include/sys/_signal.h
@@ -111,7 +111,7 @@ typedef unsigned long sigset_t;
 extern int pause(void);
 extern int raise(int signo);
 extern int kill(pid_t pid, int signo);
-extern unsigned long alarm(unsigned long sec);
+extern unsigned long alarm(unsigned sec);
 extern sigfunc_t signal(int signo, sigfunc_t func);
 
 #define SIG_BLOCK   (1)
